pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "k8s-iam-isolation"
version = "0.1.0"
description = "A CLI tool to restrict AWS IAM users, groups, and roles to Kubernetes namespaces"
authors = [
  { name = "Hooshmand k", email = "hooshmand.k@gmail.com" }
  ]
license = { text = "GNU GPL v3" }
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "click",
    "boto3",
    "kubernetes",
    "pyyaml"
]

[project.scripts]
k8s-iam = "k8s_iam_isolation.cli:cli"


[project.urls]
Homepage = "https://github.com/hooshmand/k8s-iam-isolation"
Issues = "https://github.com/hooshmand/k8s-iam-isolation/issues"

[tool.setuptools]
packages = ["k8s_iam_isolation"]
